Rename setup to locations
