+++ Assets/Python/CvEventManager.py
@@ -13,7 +13,7 @@ import CvAdvisorUtils
 from PyUtils import percentage_chance, rand, choice
 from RFCUtils import getBaseUnit, getUniqueUnit
 import Modifiers
-import Setup
+import Locations
 import random
 from Consts import MessageData
 from CoreData import civilizations, civilization
@@ -43,7 +43,7 @@ class CvEventManager:
 
     def __init__(self):
         #################### ON EVENT MAP ######################
-        Setup.setup()
+        Locations.setup()
         Modifiers.setup()
 
         self.bCtrl = False
