+++ Assets/Python/CvRFCEventHandler.py
@@ -32,7 +32,7 @@ import Stability
 import Plague
 import Crusades
 import Companies
-import Setup
+import Locations
 import Province
 import Mercenaries
 
@@ -198,7 +198,7 @@ class CvRFCEventHandler:
 
     def onGameStart(self, argsList):
         "Called at the start of the game"
-        Setup.init_values()
+        Locations.init_values()
         data.setup()
         self.province.setup()
         self.rnf.setup()
@@ -852,7 +852,7 @@ class CvRFCEventHandler:
     # This method creates a new instance of the MercenaryUtils class to be used later
     def onLoadGame(self, argsList):
         data.load()  # edead: load & unpickle script data
-        Setup.init_values()
+        Locations.init_values()
 
     # This method will redraw the main interface once a unit is promoted. This way the
     # gold/turn information will be updated.
